% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countlevels.R
\name{countlevels}
\alias{countlevels}
\title{Find out if your columns should be factors or not}
\usage{
countlevels(d)
}
\arguments{
\item{d}{A data.frame}
}
\value{
data.frame
}
\description{
Take a data.frame and check each column.  All columns that are factors
will return the column name, and whether they should be considered a
factor.  A factor should not be considered a factor when each
level is unique.
}
\examples{
countlevels(iris)
}

