% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorchars.R
\name{factorchars}
\alias{factorchars}
\title{Take a column in a data.frame and coerce it from factors to characters.}
\usage{
factorchars(q, r)
}
\arguments{
\item{q}{Your data.frame}

\item{r}{The column you wish to convert from factors to characters (as a character, like "r")}
}
\value{
data.frame
}
\description{
Take a column in a data.frame and coerce it from factors to characters.
}
\examples{
factorchars(iris, "Species")
}

