% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guelph-package.R
\docType{package}
\name{guelph}
\alias{guelph}
\alias{guelph-package}
\title{guelph.
guelph: it's what's for dinner}
\description{
Excerpt from Jenny:
Factors have driven people to extreme measures, like ordering custom
conference ribbons and laptop stickers to express how HELLNO we feel about
stringsAsFactors. And yet, sometimes you need them. Can they be made less
maddening? Let's find out.
}

