% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countlevels.R
\name{countlevels}
\alias{countlevels}
\title{Find out if your columns should be factors or not}
\usage{
countlevels(d)
}
\arguments{
\item{d}{Your dataframe}
}
\value{
data.frame
}
\description{
Find out if your columns should be factors or not
}
\examples{
countlevels(iris)
}

