% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorchars.R
\name{factorchars}
\alias{factorchars}
\title{Coerce a factor to character}
\usage{
factorchars(q, r)
}
\arguments{
\item{q}{Your list}

\item{r}{The column you wish to convert from factors to characters (as a character, like "r")}
}
\value{
list
}
\description{
Take a factor from a list and coerce it to a character
}
\examples{
factorchars(iris, "Species")
}

