% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guelph-package.r
\docType{package}
\name{guelph}
\alias{guelph}
\alias{guelph-package}
\title{guelph.
guelph: it's what's for dinner
This package is all about factor management!  Don't you hate factors?
Maybe, but they're a necessary evil. Let's make them a bit easier to work with.}
\description{
Excerpt from Jenny:
Factors have driven people to extreme measures, like ordering custom
conference ribbons and laptop stickers to express how HELLNO we feel about
stringsAsFactors. And yet, sometimes you need them. Can they be made less
maddening? Let's find out.
}

